Move sessionStore shims in to it's own file
